+++ src/docker/mqtt/start.sh
@@ -10,7 +10,7 @@ chown mosquitto:mosquitto /password_file
 
 /usr/sbin/mosquitto -c /mosquitto/config/mosquitto.conf &
 
-sleep 2
+sleep 5
 
 /app/mqtt-publisher &